chat_with_pdfs_app/requirements.txt
streamlit==1.18.1
PyPDF2==3.0.1
python-dotenv==1.0.0
langchain==0.0.184
langchain-community
huggingface-hub==0.14.1
sentence-transformers==2.2.2
torch>=1.9.0
transformers>=4.21.0
numpy>=1.21.0
scikit-learn>=1.0.0
faiss-cpu==1.7.4
altair==4
tiktoken==0.5.1
openai==0.27.6